+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 - 
 ### Fixed:
 - Fixed multi attributes
+- Fixed max health attribute
 ### Changed:
 -